Introduce generator for internal XML
